Multiple improvements
When a user views a post, if the post belongs to him, he get the option
to delete and if the post is someone elses, then he can report it.
On the posts page too, the user gets to report others posts and delete
his posts.
On the admins page, now there is the option to change an existing admin
to the captain and there is the option to remove the existing captain.